requirements.txt
appdirs==1.4.4
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
cssselect==1.2.0
fake-useragent==2.0.1
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
lxml==5.3.0
lxml_html_clean==0.4.1
numpy==2.2.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parse==1.20.2
pyee==11.1.1
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-html==0.10.0
schedule==1.2.2
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
w3lib==2.2.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
zipp==3.21.0
